Sign up, sign in, dashbboard and new board views
Create the following views:
* sign up
* sign in
* dashboard
* new board

however, it's is required to redirect a user to the sign in page
after the token has expired. Additionally, I need to handle
the result type properly. Moreover, after creating a new board,
it should redirect a user to that newly created board instead of
going to the dashboard.